requirements.txt
praw~=7.8.1
tweepy~=4.15.0
prawcore~=2.4.0
atproto~=0.0.58
python-dotenv~=1.0.1
requests~=2.32.3